Since Dashboard and Profile pages are children of Tabs page, they were moved to this folder
