+++ src/app/tabs/profile/profile.module.ts
